Move bindings into servicecatalog/client
