+++ pkg/binding/bind_cmd.go
@@ -6,6 +6,7 @@ import (
 	"github.com/Azure/service-catalog-cli/pkg/command"
 	"github.com/Azure/service-catalog-cli/pkg/output"
 	"github.com/Azure/service-catalog-cli/pkg/parameters"
+	"github.com/Azure/service-catalog-cli/pkg/service-catalog/client"
 	"github.com/spf13/cobra"
 )
 
@@ -68,12 +69,6 @@ func (c *bindCmd) run(args []string) error {
 	}
 	c.instanceName = args[0]
 
-	// Manually defaulting the name of the binding
-	// I'm not doing the same for the secret since the API handles defaulting that value.
-	if c.bindingName == "" {
-		c.bindingName = c.instanceName
-	}
-
 	params, err := parameters.ParseVariableAssignments(c.params)
 	if err != nil {
 		return fmt.Errorf("invalid --param value (%s)", err)
@@ -88,7 +83,7 @@ func (c *bindCmd) run(args []string) error {
 }
 
 func (c *bindCmd) bind(params map[string]string, secrets map[string]string) error {
-	binding, err := bind(c.Client, c.ns, c.bindingName, c.instanceName, c.secretName, params, secrets)
+	binding, err := client.Bind(c.Client, c.ns, c.bindingName, c.instanceName, c.secretName, params, secrets)
 	if err != nil {
 		return err
 	}
