+++ pkg/service-catalog/client/binding.go
@@ -1,19 +1,18 @@
-package binding
+package client
 
 import (
 	"fmt"
 	"strings"
 	"sync"
 
-	"github.com/Azure/service-catalog-cli/pkg/client"
-	traverse "github.com/Azure/service-catalog-cli/pkg/traverse"
+	"github.com/Azure/service-catalog-cli/pkg/traverse"
 	"github.com/hashicorp/go-multierror"
 	"github.com/kubernetes-incubator/service-catalog/pkg/apis/servicecatalog/v1beta1"
 	"github.com/kubernetes-incubator/service-catalog/pkg/client/clientset_generated/clientset"
 	"k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
-func retrieveAll(cl *clientset.Clientset, ns string) (*v1beta1.ServiceBindingList, error) {
+func RetrieveBindings(cl *clientset.Clientset, ns string) (*v1beta1.ServiceBindingList, error) {
 	bindings, err := cl.ServicecatalogV1beta1().ServiceBindings(ns).List(v1.ListOptions{})
 	if err != nil {
 		return nil, fmt.Errorf("unable to list bindings in %s (%s)", ns, err)
@@ -22,7 +21,7 @@ func retrieveAll(cl *clientset.Clientset, ns string) (*v1beta1.ServiceBindingLis
 	return bindings, nil
 }
 
-func retrieveByName(cl *clientset.Clientset, ns, name string) (*v1beta1.ServiceBinding, error) {
+func RetrieveBinding(cl *clientset.Clientset, ns, name string) (*v1beta1.ServiceBinding, error) {
 	binding, err := cl.ServicecatalogV1beta1().ServiceBindings(ns).Get(name, v1.GetOptions{})
 	if err != nil {
 		return nil, fmt.Errorf("unable to get binding '%s.%s' (%+v)", ns, name, err)
@@ -30,8 +29,15 @@ func retrieveByName(cl *clientset.Clientset, ns, name string) (*v1beta1.ServiceB
 	return binding, nil
 }
 
-func bind(cl *clientset.Clientset, namespace, bindingName, instanceName, secretName string,
+func Bind(cl *clientset.Clientset, namespace, bindingName, instanceName, secretName string,
 	params map[string]string, secrets map[string]string) (*v1beta1.ServiceBinding, error) {
+
+	// Manually defaulting the name of the binding
+	// I'm not doing the same for the secret since the API handles defaulting that value.
+	if bindingName == "" {
+		bindingName = instanceName
+	}
+
 	request := &v1beta1.ServiceBinding{
 		ObjectMeta: v1.ObjectMeta{
 			Name:      bindingName,
@@ -42,8 +48,8 @@ func bind(cl *clientset.Clientset, namespace, bindingName, instanceName, secretN
 				Name: instanceName,
 			},
 			SecretName:     secretName,
-			Parameters:     client.BuildParameters(params),
-			ParametersFrom: client.BuildParametersFrom(secrets),
+			Parameters:     BuildParameters(params),
+			ParametersFrom: BuildParametersFrom(secrets),
 		},
 	}
 
@@ -55,7 +61,7 @@ func bind(cl *clientset.Clientset, namespace, bindingName, instanceName, secretN
 	return result, nil
 }
 
-func unbind(cl *clientset.Clientset, ns, instanceName string) error {
+func Unbind(cl *clientset.Clientset, ns, instanceName string) error {
 	instance := &v1beta1.ServiceInstance{
 		ObjectMeta: v1.ObjectMeta{
 			Namespace: ns,
@@ -73,7 +79,7 @@ func unbind(cl *clientset.Clientset, ns, instanceName string) error {
 		g.Add(1)
 		go func(binding v1beta1.ServiceBinding) {
 			defer g.Done()
-			errs <- deleteBinding(cl, binding.Namespace, binding.Name)
+			errs <- DeleteBinding(cl, binding.Namespace, binding.Name)
 		}(binding)
 	}
 
@@ -93,7 +99,7 @@ func unbind(cl *clientset.Clientset, ns, instanceName string) error {
 	return bindErr.ErrorOrNil()
 }
 
-func deleteBinding(cl *clientset.Clientset, ns, bindingName string) error {
+func DeleteBinding(cl *clientset.Clientset, ns, bindingName string) error {
 	err := cl.ServicecatalogV1beta1().ServiceBindings(ns).Delete(bindingName, &v1.DeleteOptions{})
 	if err != nil {
 		return fmt.Errorf("remove binding %s/%s failed (%s)", ns, bindingName, err)
